+++ pyluos/modules/gpio.py
@@ -71,3 +71,24 @@ class DigitalOutputPin(object):
             self.alias,
             'high' if self.is_high() else 'low'
         )
+
+
+class Pwm(object):
+    def __init__(self, alias, delegate, default=0.0, min=0.0, max=1.0):
+        self.alias = alias
+        self._delegate = delegate
+        self._val = default
+        self._min, self._max = min, max
+
+    @property
+    def duty_cycle(self):
+        return self._val
+
+    @duty_cycle.setter
+    def duty_cycle(self, duty):
+        self._push(min(max(duty, self._min), self._max))
+
+    def _push(self, val):
+        if self._val != val:
+            self._val = val
+            self._delegate._push_value(self.alias, self._val)